refactoring, reworking algorithm and tests.  Done.
Only things left might be to consider case where user makes mistakes
and is stuck in the midpoint part of the algorithm...will take a really
long time to get where they want to go that way.  But, edge case for
another day, and, they can always type "end" and start over.  That might
be a pretty extreme edge.